refactor: reorganize ProtectedRoute and UserMenu components for improved structure
